requirements.txt
discord.py==2.5.2
aiomysql==0.2.0
python-dotenv==1.1.0
